requirements.txt
pyTelegramBotAPI==3.6.6
requests==2.7.0
urllib3==1.25.6
beautifulsoup4==4.8.1
psycopg2-binary==2.9.5
certifi
